requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
websockets==12.0
python-dotenv==1.0.1
